src/requirements.txt
pandas==2.1.2
weaviate-client==3.25.2
python-dotenv==1.0.0
streamlit==1.29.0
